requirements.txt
crewai==0.102.0
fastapi==0.115.8
neo4j==5.28.1
networkx==3.4.2
numpy==2.2.2
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
python_igraph==0.11.8
Requests==2.32.3
scipy==1.15.1
setuptools==69.2.0
tqdm==4.67.1
transformers==4.48.3
uvicorn==0.34.0
psycopg2-binary-2.9.10